Rename documentation folder
